derive: split generator implementation into 3 files
3000 lines is much too many.
